[Fix] #572 - 코드 리뷰 반영
카테고리 태그 UIView -> UILabel로 변경
